[TASK] Move email service to static MailUtility
