+++ Classes/Service/NotificationService.php
@@ -14,6 +14,7 @@ namespace DERHANSEN\SfEventMgt\Service;
  * The TYPO3 project - inspiring people to share!
  */
 
+use DERHANSEN\SfEventMgt\Utility\MailUtility;
 use DERHANSEN\SfEventMgt\Utility\MessageType;
 use DERHANSEN\SfEventMgt\Utility\MessageRecipient;
 use \TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -40,14 +41,6 @@ class NotificationService
      * @inject
      */
     protected $registrationRepository = null;
-
-    /**
-     * Email Service
-     *
-     * @var \DERHANSEN\SfEventMgt\Service\EmailService
-     * @inject
-     */
-    protected $emailService;
 
     /**
      * Hash Service
@@ -185,7 +178,7 @@ class NotificationService
                 $type,
                 MessageRecipient::USER
             );
-            return $this->emailService->sendEmailMessage(
+            return MailUtility::sendEmailMessage(
                 $settings['notification']['senderEmail'],
                 $registration->getEmail(),
                 $subject,
@@ -278,7 +271,7 @@ class NotificationService
         if ($event->getNotifyAdmin()) {
             $adminEmailArr = GeneralUtility::trimExplode(',', $settings['notification']['adminEmail'], true);
             foreach ($adminEmailArr as $adminEmail) {
-                $allEmailsSent = $allEmailsSent && $this->emailService->sendEmailMessage(
+                $allEmailsSent = $allEmailsSent && MailUtility::sendEmailMessage(
                     $settings['notification']['senderEmail'],
                     $adminEmail,
                     $subject,
@@ -289,7 +282,7 @@ class NotificationService
             }
         }
         if ($event->getNotifyOrganisator() && $event->getOrganisator()) {
-            $allEmailsSent = $allEmailsSent && $this->emailService->sendEmailMessage(
+            $allEmailsSent = $allEmailsSent && MailUtility::sendEmailMessage(
                 $settings['notification']['senderEmail'],
                 $event->getOrganisator()->getEmail(),
                 $subject,
